% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{add_line}
\alias{add_line}
\title{Add light}
\usage{
add_line(lines, line)
}
\arguments{
\item{obj_model}{}
}
\value{
Rasterized image.
}
\description{
Add light
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
