% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plane_uv_mesh.R
\name{add_sphere_uv_mesh}
\alias{add_sphere_uv_mesh}
\title{Add Sphere UV Mapping to Mesh}
\usage{
add_sphere_uv_mesh(mesh, origin = c(0, 0, 0), override_existing = FALSE)
}
\arguments{
\item{mesh}{The mesh to which the UV mapping will be applied.}

\item{origin}{Default \code{c(0, 0, 0)}. A vector specifying the origin to
apply spherical UV coordinates.}

\item{override_existing}{Default \code{FALSE}. Specifies whether existing UV
coordinates should be overridden.}
}
\value{
Modified mesh with added UV mapping.
}
\description{
Applies a planar UV mapping to a mesh based on a spherical
direction from the origin.
}
\examples{
if(run_documentation()) {
#Let's construct a mesh from the volcano dataset

}
}
