% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayvertex_classes.R
\name{cat_color}
\alias{cat_color}
\title{Print out the color}
\usage{
cat_color(color, var_name, default = NA, intensity = 1, spacer = "")
}
\value{
none
}
\description{
Print out the color
}
\keyword{internal}
