% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shape.R
\name{change_material}
\alias{change_material}
\title{Change Material

Add shape to a mesh}
\usage{
change_material(
  mesh,
  diffuse = NULL,
  ambient = NULL,
  specular = NULL,
  transmittance = NULL,
  emission = NULL,
  shininess = NULL,
  ior = NULL,
  dissolve = NULL,
  illum = NULL,
  texture_location = NULL,
  normal_texture_location = NULL,
  specular_texture_location = NULL,
  ambient_texture_location = NULL,
  emissive_texture_location = NULL,
  diffuse_intensity = NULL,
  specular_intensity = NULL,
  emission_intensity = NULL,
  culling = NULL,
  type = NULL
)
}
\arguments{
\item{obj_model}{}
}
\value{
Rasterized image.
}
\description{
Change Material

Add shape to a mesh
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
