% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{color_lines}
\alias{color_lines}
\title{Color Lines}
\usage{
color_lines(lines, color = "white")
}
\arguments{
\item{lines}{The line scene.}

\item{color}{Default `white`. The color to convert the lines to.}
}
\value{
Colored line matrix.
}
\description{
Color Lines
}
\examples{
#Generate a cube out of lines
\donttest{
cube_outline = generate_line(start = c(-1, -1, -1), end = c(-1, -1, 1)) \%>\%
  add_lines(generate_line(start = c(-1, -1, -1), end = c(-1, 1, -1))) \%>\%
  add_lines(generate_line(start = c(-1, -1, -1), end = c(1, -1, -1))) \%>\%
  add_lines(generate_line(start = c(-1, -1, 1), end = c(-1, 1, 1))) \%>\%
  add_lines(generate_line(start = c(-1, -1, 1), end = c(1, -1, 1))) \%>\%
  add_lines(generate_line(start = c(-1, 1, 1), end = c(-1, 1, -1))) \%>\%
  add_lines(generate_line(start = c(-1, 1, 1), end = c(1, 1, 1))) \%>\%
  add_lines(generate_line(start = c(1, 1, -1), end = c(1, -1, -1))) \%>\%
  add_lines(generate_line(start = c(1, 1, -1), end = c(1, 1, 1))) \%>\%
  add_lines(generate_line(start = c(1, -1, -1), end = c(1, -1, 1))) \%>\%
  add_lines(generate_line(start = c(1, -1, 1), end = c(1, 1, 1))) \%>\%
  add_lines(generate_line(start = c(-1, 1, -1), end = c(1, 1, -1)))
  
cube_outline \%>\% 
  color_lines(color="red") \%>\% 
  rasterize_lines()
  }
}
