% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_mesh.R
\name{construct_mesh}
\alias{construct_mesh}
\title{Manually construct a mesh}
\usage{
construct_mesh(
  vertices,
  indices,
  normals = NULL,
  norm_indices = NULL,
  texcoords = NULL,
  tex_indices = NULL,
  material = material_list()
)
}
\arguments{
\item{vertices}{Nx3 matrix of vertex coordinates..}

\item{indices}{Nx3 integer matrix, where each row defines a triangle using the 
vertices defined in `vertices`.}

\item{normals}{Default `NULL`.  Nx3 matrix of normals.}

\item{norm_indices}{Nx3 integer matrix, where each row defines the normal for a vertex using the 
normals defined in `normals` for the corresponding triangle in `indices`.
Required to be the same number of rows as `indices`.}

\item{texcoords}{Default `NULL`. Nx2 matrix of texture coordinates.}

\item{tex_indices}{Nx3 integer matrix, where each row defines the texture coordinates for a triangle
using the tex coords defined in `texcoors` for the corresponding triangle in `indices`.
Required to be the same number of rows as `indices`.}

\item{material}{Default `material_list()` (default values). Specify the material of the object.}
}
\value{
Rasterized image.
}
\description{
Manually construct a mesh
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
