% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayvertex_classes.R
\name{convert_rgb_to_ansi}
\alias{convert_rgb_to_ansi}
\title{Convert RGB to ANSI Color}
\usage{
convert_rgb_to_ansi(color)
}
\arguments{
\item{color}{Length=3 numeric vector.}
}
\value{
ANSI color code as a string.
}
\description{
Convert RGB to ANSI Color
}
\keyword{internal}
