% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plane_uv_mesh.R
\name{cross_prod}
\alias{cross_prod}
\title{Cross Product}
\usage{
cross_prod(x, y)
}
\arguments{
\item{x}{A numeric vector representing the first 3D vector.}

\item{y}{A numeric vector representing the second 3D vector.}
}
\value{
A numeric vector representing the cross product of \code{x} and \code{y}.
}
\description{
Computes the cross product of two 3-dimensional vectors.
}
\keyword{internal}
