% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{cube_mesh}
\alias{cube_mesh}
\title{Cube 3D Model}
\usage{
cube_mesh(
  position = c(0, 0, 0),
  scale = c(1, 1, 1),
  angle = c(0, 0, 0),
  pivot_point = c(0, 0, 0),
  order_rotation = c(1, 2, 3)
)
}
\arguments{
\item{position}{Default `c(0,0,0)`. Position of the mesh.}

\item{scale}{Default `c(1,1,1)`. Scale of the mesh. Can also be a single numeric value scaling all axes uniformly.}

\item{angle}{Default `c(0,0,0)`. Angle to rotate the mesh.}

\item{pivot_point}{Default `c(0,0,0)`. Point around which to rotate the mesh.}

\item{order_rotation}{Default `c(1,2,3)`. Order to rotate the axes.}
}
\value{
File location of the R.obj file (saved with a .txt extension)
}
\description{
3D obj model of the letter R
}
\examples{
#Load and render the included example R object file.
}
