% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_obj.R
\name{cube_obj}
\alias{cube_obj}
\title{Cube 3D Model}
\usage{
cube_obj()
}
\value{
File location of the R.obj file (saved with a .txt extension)
}
\description{
3D obj model of the letter R
}
\examples{
#Load and render the included example R object file.
}
