% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{cylinder_mesh}
\alias{cylinder_mesh}
\title{Cube 3D Model}
\usage{
cylinder_mesh(
  position = c(0, 0, 0),
  radius = 0.5,
  length = 1,
  angle = c(0, 0, 0),
  pivot_point = c(0, 0, 0),
  order_rotation = c(1, 2, 3),
  material = material()
)
}
\arguments{
\item{position}{Default `c(0,0,0)`. Position of the mesh.}

\item{radius}{Default `0.5`. Radius of the cylinder.}

\item{length}{Default `1`. Length of the cylinder.}

\item{angle}{Default `c(0,0,0)`. Angle to rotate the mesh.}

\item{pivot_point}{Default `c(0,0,0)`. Point around which to rotate the mesh.}

\item{order_rotation}{Default `c(1,2,3)`. Order to rotate the axes.}
}
\value{
File location of the R.obj file (saved with a .txt extension)
}
\description{
Cube 3D Model
}
