% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lights.R
\name{directional_light}
\alias{directional_light}
\title{Generate Directional Lights}
\usage{
directional_light(direction = c(0, 1, 0), color = "white", intensity = 1)
}
\arguments{
\item{obj_model}{A two-dimensional matrix, where each entry in the matrix is the elevation at that point. All points are assumed to be evenly spaced.}

\item{width}{Default `400`. Width of the rendered image.}

\item{height}{Default `400`. Width of the rendered image.}

\item{fov}{Default `20`. Width of the rendered image.}

\item{lookfrom}{Default `c(0,0,10)`. Camera location.}

\item{lookat}{Default `c(0,0,0)`. Camera focal position.}
}
\value{
Rasterized image.
}
\description{
Generate Directional Lights
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
