% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displace_mesh.R
\name{displace_mesh}
\alias{displace_mesh}
\title{Displace a Mesh}
\usage{
displace_mesh(
  mesh,
  displacement_texture,
  displacement_scale = 1,
  displacement_vector = FALSE,
  id = NA
)
}
\arguments{
\item{mesh}{The mesh.}

\item{displacement_texture}{Image or matrix/array that will be used to displace the mesh}

\item{displacement_scale}{Default \code{1}. Intensity of the displacement effect.
Higher values result in greater displacement.}

\item{displacement_vector}{Default \code{FALSE}. Whether to use vector displacement.
If \code{TRUE}, the displacement texture is interpreted as providing a 3D displacement vector.
Otherwise, the texture is interpreted as providing a scalar displacement.}

\item{id}{Default \code{NA} (all shapes). The shape index to have new normals calculated.}
}
\value{
raymesh object
}
\description{
Displace a Mesh
}
\examples{
if(run_documentation()) {
 #Let's construct a mesh from the volcano dataset
}
}
