% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_orientation_mesh.R
\name{flip_orientation_mesh}
\alias{flip_orientation_mesh}
\title{Flip Orientation}
\usage{
flip_orientation_mesh(mesh)
}
\arguments{
\item{mesh}{The mesh to swap orientations.}
}
\value{
Mesh with flipped vertex orientation
}
\description{
Flip Orientation
}
\examples{
# Flip a mesh 
if(run_documentation()) {
sphere_mesh(position=c(-1,0,0)) |> 
  add_shape(flip_orientation_mesh(sphere_mesh(position=c(1,0,0)))) |> 
  rasterize_scene(debug="normals",fov=30)
}
}
