% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayvertex_classes.R
\name{format_pillar_shp}
\alias{format_pillar_shp}
\title{Format ray_shape for pillar}
\usage{
format_pillar_shp(x)
}
\arguments{
\item{x}{A ray_shape object.}
}
\value{
A formatted character vector.
}
\description{
Format ray_shape for pillar
}
\keyword{internal}
