% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{generate_cornell_mesh}
\alias{generate_cornell_mesh}
\title{Rasterize an OBJ file}
\usage{
generate_cornell_mesh(
  leftcolor = "#1f7326",
  rightcolor = "#a60d0d",
  roomcolor = "#bababa",
  ceiling = TRUE,
  light = TRUE,
  simulate_diffuse_light = TRUE
)
}
\arguments{
\item{leftcolor}{Default `#1f7326` (green).}

\item{rightcolor}{Default `#a60d0d` (red).}

\item{roomcolor}{Default `#bababa` (light grey).}

\item{ceiling}{Default `TRUE`. Whether to render the ceiling.}

\item{light}{Default `TRUE`. Whether to render a point light near the ceiling.}

\item{simulate_diffuse_light}{Default `TRUE`. Whether place point lights in the scene to 
simulate the first diffuse bounce of light.}
}
\value{
Rasterized image.
}
\description{
Rasterize an OBJ file
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
