% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shape.R
\name{generate_rot_matrix}
\alias{generate_rot_matrix}
\title{Generate Rotation Matrix}
\usage{
generate_rot_matrix(angle, order_rotation)

generate_rot_matrix(angle, order_rotation)
}
\arguments{
\item{angle}{The angle}

\item{order_rotation}{Default `c(1,2,3)`.}
}
\value{
Matrix

Matrix
}
\description{
Generate Rotation Matrix
}
\keyword{#Here}
\keyword{`montereybay`}
\keyword{a}
\keyword{ambient}
\keyword{elevation}
\keyword{internal}
\keyword{map}
\keyword{map.}
\keyword{occlusion}
\keyword{of}
\keyword{produce}
\keyword{the}
\keyword{we}
