% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plane_uv_mesh.R
\name{generate_rotation_matrix_from_direction}
\alias{generate_rotation_matrix_from_direction}
\title{Generate Rotation Matrix from Direction}
\usage{
generate_rotation_matrix_from_direction(
  direction = c(0, 1, 0),
  up = c(0, 1, 0)
)
}
\arguments{
\item{direction}{Default \code{c(0, 1, 0)}. A 3D vector representing the direction.
The function normalizes this vector.}

\item{up}{Default \code{c(0, 1, 0)}. A 3D vector representing the up direction.}
}
\value{
A 3x3 rotation matrix.
}
\description{
Internal function to generate a rotation matrix that aligns the
Y-axis with a given direction vector. It uses the Rodrigues' rotation formula.
}
\keyword{internal}
