% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shape.R
\name{generate_toon_outline}
\alias{generate_toon_outline}
\title{Add Outline}
\usage{
generate_toon_outline(single_obj, material, scale = 1)
}
\arguments{
\item{angle}{The angle}

\item{order_rotation}{Default `c(1,2,3)`.}
}
\value{
Matrix
}
\description{
Add Outline
}
\keyword{#Here}
\keyword{`montereybay`}
\keyword{a}
\keyword{ambient}
\keyword{elevation}
\keyword{internal}
\keyword{map}
\keyword{map.}
\keyword{occlusion}
\keyword{of}
\keyword{produce}
\keyword{the}
\keyword{we}
