% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{get_file_type}
\alias{get_file_type}
\title{Check Filename}
\usage{
get_file_type(file)
}
\arguments{
\item{file}{Filename to be checked}
}
\value{
Flipped matrix
}
\description{
Check Filename
}
\examples{
#Fake example
}
\keyword{internal}
