% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_mesh_functions.R
\name{get_mesh_bbox}
\alias{get_mesh_bbox}
\title{Get Mesh Bounding Box}
\usage{
get_mesh_bbox(mesh)
}
\arguments{
\item{mesh}{The mesh object.}
}
\value{
2x3 numeric matrix
}
\description{
Calculates the bounding box of a mesh
}
\examples{
if(run_documentation()) {
#Calculates the center of the mesh
get_mesh_bbox(generate_cornell_mesh())
}
}
