% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_mesh_functions.R
\name{get_mesh_center}
\alias{get_mesh_center}
\title{Get Mesh Center}
\usage{
get_mesh_center(mesh)
}
\arguments{
\item{mesh}{The mesh object.}
}
\value{
Length-3 numeric vector
}
\description{
Calculates the coordinates of the center of a mesh
}
\examples{
if(run_documentation()) {
#Calculates the center of the mesh
get_scene_center(generate_cornell_mesh())
}
}
