% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_scene.R
\name{merge_scene}
\alias{merge_scene}
\title{Merge scene}
\usage{
merge_scene(old_scene, flatten_materials = TRUE)
}
\arguments{
\item{scene}{}
}
\value{
Merged scene
}
\description{
Merge the shapes to one
}
\keyword{internal}
