% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shape.R
\name{merge_shapes}
\alias{merge_shapes}
\title{Add Shape

Add shape to a mesh}
\usage{
merge_shapes(mesh)
}
\arguments{
\item{obj_model}{}
}
\value{
Rasterized image.
}
\description{
Add Shape

Add shape to a mesh
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
