% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plane_uv_mesh.R
\name{normalize}
\alias{normalize}
\title{Normalize Vector}
\usage{
normalize(v)
}
\arguments{
\item{v}{A numeric vector}
}
\value{
Normilzed vector
}
\description{
Computes a normalized vector
}
\keyword{internal}
