% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{obj_mesh}
\alias{obj_mesh}
\title{Rasterize an OBJ file}
\usage{
obj_mesh(
  filename,
  position = c(0, 0, 0),
  scale = c(1, 1, 1),
  angle = c(0, 0, 0),
  pivot_point = c(0, 0, 0),
  order_rotation = c(1, 2, 3),
  materialspath = NULL
)
}
\arguments{
\item{filename}{OBJ filename.}

\item{position}{Default `c(0,0,0)`. Position of the mesh.}

\item{scale}{Default `c(1,1,1)`. Scale of the mesh. Can also be a single numeric value scaling all axes uniformly.}

\item{angle}{Default `c(0,0,0)`. Angle to rotate the mesh.}

\item{pivot_point}{Default `c(0,0,0)`. Point around which to rotate the mesh.}

\item{order_rotation}{Default `c(1,2,3)`. Order to rotate the axes.}

\item{materialspath}{Default `NULL`. Path to the MTL file, if different from the OBJ file.}
}
\value{
Rasterized image.
}
\description{
Rasterize an OBJ file
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
