% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayvertex_classes.R
\name{pillar_shaft.ray_shape_list}
\alias{pillar_shaft.ray_shape_list}
\title{Pillar shaft for ray_shape_list}
\usage{
pillar_shaft.ray_shape_list(x, ...)
}
\arguments{
\item{x}{A ray_shape_list object.}

\item{...}{Additional arguments (unused).}
}
\value{
A pillar shaft object.
}
\description{
Pillar shaft for ray_shape_list
}
\keyword{internal}
