% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{plane_xy_mesh}
\alias{plane_xy_mesh}
\title{Cube 3D Model}
\usage{
plane_xy_mesh(position = c(0, 0, 0))
}
\value{
File location of the R.obj file (saved with a .txt extension)
}
\description{
3D obj model of the letter R
}
