% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{plane_yz_mesh}
\alias{plane_yz_mesh}
\title{Cube 3D Model}
\usage{
plane_yz_mesh(
  position = c(0, 0, 0),
  scale = c(1, 1, 1),
  angle = c(0, 0, 0),
  pivot_point = c(0, 0, 0),
  order_rotation = c(1, 2, 3)
)
}
\value{
File location of the R.obj file (saved with a .txt extension)
}
\description{
3D obj model of the letter R
}
