% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shape.R
\name{preprocess_scene}
\alias{preprocess_scene}
\title{Preprocess Scene}
\usage{
preprocess_scene(scene_list)
}
\arguments{
\item{scene_list}{The scene list.}
}
\value{
Scene list converted to proper format.
}
\description{
This function takes a ray_mesh object (which is a list that has multiple vertex/
normal/tex matrices and materials for each, as well as shapes that refer to each one) and combines them into a single set
of vertex/normal/tex matrices and materials by adjusting the index arguments in each shape.
}
\keyword{internal}
