% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayvertex_classes.R
\name{print.ray_shape}
\alias{print.ray_shape}
\title{Print method for ray_shape}
\usage{
\method{print}{ray_shape}(x, ...)
}
\arguments{
\item{x}{A ray_shape object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print method for ray_shape
}
\keyword{internal}
