% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayvertex_classes.R
\name{print.ray_vertex_data}
\alias{print.ray_vertex_data}
\title{Print method for ray_vertex_data}
\usage{
\method{print}{ray_vertex_data}(x, ...)
}
\arguments{
\item{x}{A ray_vertex_data object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print method for ray_vertex_data
}
\keyword{internal}
