% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_output.R
\name{raster_output_image}
\alias{raster_output_image}
\title{Assemble the renderer's planar channels into an oriented linear image}
\usage{
raster_output_image(imagelist)
}
\arguments{
\item{imagelist}{Native render buffers.}
}
\description{
Assemble the renderer's planar channels into an oriented linear image
}
\keyword{internal}
