% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_mesh.R
\name{rasterize_mesh}
\alias{rasterize_mesh}
\title{Rasterize an OBJ file}
\usage{
rasterize_mesh(
  vertices,
  indices = NULL,
  texcoords = NULL,
  normals = NULL,
  filename = NA,
  width = 400,
  height = 400,
  fov = 20,
  lookfrom = c(0, 0, 10),
  lookat = c(0, 0, 0),
  type = "diffuse",
  color = "darkred",
  texture_location = NA,
  normal_texture_location = NA,
  specular_texture_location = NA,
  emissive_texture_location = NA,
  parallel = TRUE,
  light_direction = c(1, 1, 1),
  ambient_color = c(0, 0, 0),
  exponent = 32,
  specular_intensity = 0.6,
  emission_intensity = 1,
  diffuse_intensity = 1,
  tangent_space_normals = FALSE,
  shadow_map = FALSE,
  calc_ambient = FALSE,
  ambient_intensity = 10,
  ambient_radius = 0.1,
  tonemap = "none",
  debug = "none",
  shadow_map_bias = 0.005
)
}
\arguments{
\item{width}{Default `400`. Width of the rendered image.}

\item{height}{Default `400`. Width of the rendered image.}

\item{fov}{Default `20`. Width of the rendered image.}

\item{lookfrom}{Default `c(0,0,10)`. Camera location.}

\item{lookat}{Default `c(0,0,0)`. Camera focal position.}

\item{obj_model}{A two-dimensional matrix, where each entry in the matrix is the elevation at that point. All points are assumed to be evenly spaced.}
}
\value{
Rasterized image.
}
\description{
Rasterize an OBJ file
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
