% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_mesh3d.R
\name{rasterize_mesh3d}
\alias{rasterize_mesh3d}
\title{Rasterize an OBJ file

If the OBJ file has an MTL file, it must be located in the same directory as the OBJ file itself.
Additionally, the textures referenced in the MTL file must be specified as a 
relative path to the OBJ file's location.}
\usage{
rasterize_mesh3d(
  mesh,
  diffuse = c(0.5, 0.5, 0.5),
  ambient = c(0, 0, 0),
  dissolve = 0,
  shape = NULL,
  ...
)
}
\arguments{
\item{mesh}{Filename of the `obj` file.}

\item{shape}{Default `NULL`. The shape to render in the OBJ mesh.}

\item{...}{Arguments to pass to `rasterize_mesh()`}
}
\value{
Rasterized image.
}
\description{
Rasterize an OBJ file

If the OBJ file has an MTL file, it must be located in the same directory as the OBJ file itself.
Additionally, the textures referenced in the MTL file must be specified as a 
relative path to the OBJ file's location.
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
