% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayvertex_classes.R
\name{ray_shape}
\alias{ray_shape}
\title{Define the ray_shape class}
\usage{
ray_shape(...)
}
\arguments{
\item{...}{Objects to be included in the ray_shape.}
}
\value{
A new ray_shape object.
}
\description{
Define the ray_shape class
}
\keyword{internal}
