% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayvertex_classes.R
\name{ray_shape_list}
\alias{ray_shape_list}
\title{Define the ray_shape_list class}
\usage{
ray_shape_list(...)
}
\arguments{
\item{...}{Objects to be included in the ray_shape_list.}
}
\value{
A new ray_shape_list object.
}
\description{
Define the ray_shape_list class
}
\keyword{internal}
