% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayvertex_classes.R
\name{ray_vertex_data}
\alias{ray_vertex_data}
\title{Define the ray_vertex_data class}
\usage{
ray_vertex_data(data = NA)
}
\arguments{
\item{data}{A matrix with 2 or 3 columns representing vertex data.}
}
\value{
A new ray_vertex_data object.
}
\description{
Define the ray_vertex_data class
}
\keyword{internal}
