% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readobj.R
\name{read_ply}
\alias{read_ply}
\title{Load an PLY file}
\usage{
read_ply(filename)
}
\value{
Rasterized image.
}
\description{
Load an PLY file
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
\keyword{internal}
