% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{rescale}
\alias{rescale}
\title{Rescale}
\usage{
rescale(vals, to = c(0, 1))
}
\arguments{
\item{vals}{Values}

\item{to}{to vales}
}
\value{
Color vector
}
\description{
Rescale
}
\examples{
#none
}
\keyword{internal}
