% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_scene_obj.R
\name{save_scene_obj}
\alias{save_scene_obj}
\title{Save scene to OBJ file}
\usage{
save_scene_obj(scene, filename, materials = TRUE, fileext = ".obj")
}
\arguments{
\item{scene}{A rayvertex scene.}

\item{filename}{The filename for the OBJ file.}

\item{materials}{Default `TRUE`. Whether to write an MTL file to specify the materials for the OBJ.}

\item{fileext}{Default `".obj"`. The file extension to add to the filename.}
}
\value{
None
}
\description{
Writes the current scene to a Wavefront OBJ file, with or without materials
}
\examples{
if(rayvertex:::run_documentation()) {
tmpfile = tempfile(fileext = ".obj")
save_scene_obj(generate_cornell_mesh(), tmpfile)
}
}
