% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{segment_mesh}
\alias{segment_mesh}
\title{Cube 3D Model}
\usage{
segment_mesh(
  start = c(0, -1, 0),
  end = c(0, 1, 0),
  radius = 0.5,
  direction = NA,
  from_center = TRUE,
  square = FALSE
)
}
\value{
File location of the R.obj file (saved with a .txt extension)
}
\description{
3D obj model of the letter R
}
