% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shape.R
\name{set_material}
\alias{set_material}
\title{Set Material

Add shape to a mesh}
\usage{
set_material(
  mesh,
  exponent = 32,
  ambient = c(0, 0, 0),
  diffuse = c(0.5, 0.5, 0.5),
  specular = c(1, 1, 1),
  transmittance = c(1, 1, 1),
  emission = c(0, 0, 0),
  shininess = 10,
  ior = 1,
  dissolve = 1,
  illum = 1,
  texture_location = "",
  normal_texture_location = "",
  specular_texture_location = "",
  ambient_texture_location = "",
  emissive_texture_location = "",
  diffuse_intensity = 1,
  specular_intensity = 0.6,
  emission_intensity = 1,
  culling = "back"
)
}
\arguments{
\item{obj_model}{}
}
\value{
Rasterized image.
}
\description{
Set Material

Add shape to a mesh
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
