% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{sphere_mesh}
\alias{sphere_mesh}
\title{Sphere 3D Model}
\usage{
sphere_mesh(
  position = c(0, 0, 0),
  scale = c(1, 1, 1),
  angle = c(0, 0, 0),
  pivot_point = c(0, 0, 0),
  order_rotation = c(1, 2, 3),
  radius = 1,
  low_poly = FALSE,
  material = material()
)
}
\arguments{
\item{position}{Default `c(0,0,0)`. Position of the mesh.}

\item{scale}{Default `c(1,1,1)`. Scale of the mesh. Can also be a single numeric value scaling all axes uniformly.}

\item{angle}{Default `c(0,0,0)`. Angle to rotate the mesh.}

\item{pivot_point}{Default `c(0,0,0)`. Point around which to rotate the mesh.}

\item{order_rotation}{Default `c(1,2,3)`. Order to rotate the axes.}

\item{radius}{Default `1`. Radius of the sphere.}

\item{low_poly}{Default `FALSE`. If `TRUE`, will use a low-poly sphere.}
}
\value{
File location of the R.obj file (saved with a .txt extension)
}
\description{
Sphere 3D Model
}
\examples{
#Generate a sphere in the cornell box.
\donttest{
generate_cornell_mesh() \%>\%
  add_shape(sphere_mesh(position = c(555/2, 555/2, 555/2), radius = 100)) \%>\%
  rasterize_mesh(width=800,height=800)
}

#Generate a shiny sphere in the cornell box
\donttest{
generate_cornell() \%>\%
  add_object(sphere(x = 555/2, y = 100, z = 555/2, radius = 100, 
                    material = microfacet(color = "gold"))) \%>\%
  render_scene(lookfrom = c(278, 278, -800) ,lookat = c(278, 278, 0), fov = 40, 
               ambient_light = FALSE, samples = 400, clamp_value = 5)
}
}
