% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{text3d_mesh}
\alias{text3d_mesh}
\title{Text Object}
\usage{
text3d_mesh(
  label,
  position = c(0, 0, 0),
  text_height = 1,
  orientation = "xy",
  color = "black",
  angle = c(0, 0, 0),
  pivot_point = c(0, 0, 0),
  order_rotation = c(1, 2, 3),
  scale = c(1, 1, 1)
)
}
\arguments{
\item{label}{Text string.}

\item{text_height}{Default `1`. Height of the text.}

\item{orientation}{Default `xy`. Orientation of the plane. Other options are `yz` and `xz`.}

\item{angle}{Default `c(0, 0, 0)`. Angle of rotation around the x, y, and z axes, applied in the order specified in `order_rotation`.}

\item{order_rotation}{Default `c(1, 2, 3)`. The order to apply the rotations, referring to "x", "y", and "z".}

\item{scale}{Default `c(1, 1, 1)`. Scale transformation in the x, y, and z directions. If this is a single value,
number, the object will be scaled uniformly.
Note: emissive objects may not currently function correctly when scaled.}

\item{x}{Default `0`. x-coordinate of the center of the label.}

\item{y}{Default `0`. y-coordinate of the center of the label.}

\item{z}{Default `0`. z-coordinate of the center of the label.}

\item{material}{Default  \code{\link{diffuse}}. The material, called from one of the material 
functions \code{\link{diffuse}}, \code{\link{metal}}, or \code{\link{dielectric}}.}

\item{velocity}{Default `c(0, 0, 0)`. Velocity of the sphere, used for motion blur.}

\item{flipped}{Default `FALSE`. Whether to flip the normals.}
}
\value{
Single row of a tibble describing the text in the scene.
}
\description{
Text Object
}
\examples{
#Generate a label in the cornell box.
\donttest{
generate_cornell_mesh() \%>\% 
  add_shape(text3d_mesh(label="Cornell Box", position=c(555/2,555/2,555/2),text_height=60)) \%>\% 
  rasterize_scene(light_info = directional_light(c(0.1,0.4,-1)))
  
#Change the orientation
generate_cornell_mesh() \%>\% 
  add_shape(text3d_mesh(label="YZ Plane", position=c(540,555/2,555/2),text_height=100,
                    orientation = "yz",angle=c(0,0,0))) \%>\% 
  add_shape(text3d_mesh(label="XY Plane", position=c(555/2,555/2,540),text_height=100,
                    orientation = "xy", angle=c(0,0,0))) \%>\% 
  add_shape(text3d_mesh(label="XZ Plane", position=c(555/2,15,555/2),text_height=100,
                    orientation = "xz")) \%>\% 
  rasterize_scene(light_info = directional_light(c(0.1,0.4,-1)))
  
#Add an label in front of a sphere
generate_cornell_mesh() \%>\% 
  add_shape(text3d_mesh(label="Cornell Box", position=c(555/2,555/2,555/2),text_height=60,
                    color="grey20")) \%>\% 
  add_shape(text3d_mesh(label="Sphere", position=c(555/2,100,100),text_height=30,
                    color="white")) \%>\% 
  add_shape(sphere_mesh(radius=100,position=c(555/2,100,555/2),
                    material=material_list(diffuse="purple",type="phong"))) \%>\%                  
  rasterize_scene(light_info = directional_light(c(0.1,0.4,-1)))
  
  
#A room full of bees
bee_scene = list()
for(i in 1:100) {
bee_scene = add_shape(bee_scene, text3d_mesh("B", position=c(20+runif(3)*525), 
                                             color="yellow", text_height = 50,
                                             angle=c(0,0,0)))
}
generate_cornell_mesh(light=T) \%>\% 
  add_shape(bee_scene) \%>\%                   
  rasterize_scene(light=directional_light(c(0,1,-1)))
}
}
