% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shape.R
\name{translate_shape}
\alias{translate_shape}
\title{Add Shape

Add shape to a mesh}
\usage{
translate_shape(mesh, position = c(0, 0, 0))
}
\arguments{
\item{obj_model}{}
}
\value{
Rasterized image.
}
\description{
Add Shape

Add shape to a mesh
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
