% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_scene.R
\name{validate_scene}
\alias{validate_scene}
\title{Validate Scene}
\usage{
validate_scene(scene)
}
\arguments{
\item{scene}{Make sure that there are no out of bounds issues and all the materials are valid}
}
\value{
Color vector
}
\description{
Validate Scene
}
\keyword{internal}
