% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayvertex_classes.R
\name{vec_ptype_abbr.ray_shape}
\alias{vec_ptype_abbr.ray_shape}
\title{Abbreviate the ptype of ray_shape}
\usage{
vec_ptype_abbr.ray_shape(x, ...)
}
\arguments{
\item{x}{A ray_shape object.}

\item{...}{Additional arguments (unused).}
}
\value{
A character vector with the abbreviation "ray_shp".
}
\description{
Abbreviate the ptype of ray_shape
}
\keyword{internal}
