% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lights.R
\name{add_light}
\alias{add_light}
\title{Add light}
\usage{
add_light(light_scene, light)
}
\arguments{
\item{obj_model}{}
}
\value{
Rasterized image.
}
\description{
Add light
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
