% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shape.R
\name{add_shape}
\alias{add_shape}
\title{Add Shape

Add shape to a mesh}
\usage{
add_shape(mesh, shape)
}
\arguments{
\item{obj_model}{}
}
\value{
Rasterized image.
}
\description{
Add Shape

Add shape to a mesh
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
