% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{arrow_mesh}
\alias{arrow_mesh}
\title{Cube 3D Model}
\usage{
arrow_mesh(
  start = c(0, 0, 0),
  end = c(0, 1, 0),
  scale = c(1, 1, 1),
  radius_head = 0.5,
  radius_tail = 0.25,
  tail_proportion = 0.5,
  direction = NA,
  from_center = TRUE
)
}
\value{
File location of the R.obj file (saved with a .txt extension)
}
\description{
3D obj model of the letter R
}
