% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{arrow_mesh}
\alias{arrow_mesh}
\title{Cube 3D Model}
\usage{
arrow_mesh(
  start = c(0, 0, 0),
  end = c(0, 1, 0),
  radius_top = 0.5,
  radius_tail = 0.25,
  tail_proportion = 0.5,
  direction = NA,
  from_center = TRUE,
  material = material()
)
}
\arguments{
\item{start}{Default `c(0, 0, 0)`. Base of the arrow, specifying `x`, `y`, `z`.}

\item{end}{Default `c(0, 1, 0)`. Tip of the arrow, specifying `x`, `y`, `z`.}

\item{radius_top}{Default `0.5`. Radius of the top of the arrow.}

\item{radius_tail}{Default `0.2`.  Radius of the tail of the arrow.}

\item{tail_proportion}{Default `0.5`. Proportion of the arrow that is the tail.}

\item{direction}{Default `NA`. Alternative to `start` and `end`, specify the direction (via 
a length-3 vector) of the arrow. Arrow will be centered at `start`, and the length will be
determined by the magnitude of the direction vector.}

\item{from_center}{Default `TRUE`. If orientation specified via `direction`, setting this argument
to `FALSE` will make `start` specify the bottom of the cone, instead of the middle.}
}
\value{
File location of the R.obj file (saved with a .txt extension)
}
\description{
Cube 3D Model
}
