% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{cone_mesh}
\alias{cone_mesh}
\title{Cube 3D Model}
\usage{
cone_mesh(
  position = c(0, 0, 0),
  scale = c(1, 1, 1),
  angle = c(0, 0, 0),
  radius_head = 0.5,
  radius_tail = 0.25,
  pivot_point = c(0, 0, 0),
  order_rotation = c(1, 2, 3)
)
}
\value{
File location of the R.obj file (saved with a .txt extension)
}
\description{
3D obj model of the letter R
}
