% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayvertex_classes.R
\name{convert_rgb_to_ansi}
\alias{convert_rgb_to_ansi}
\title{Convert RGB to ANSI Color}
\usage{
convert_rgb_to_ansi(color)
}
\arguments{
\item{r}{Default \code{1}. Red component in the range \link{0, 1}.}

\item{g}{Default \code{1}. Green component in the range \link{0, 1}.}

\item{b}{Default \code{1}. Blue component in the range \link{0, 1}.}
}
\value{
ANSI color code as a string.
}
\description{
Convert RGB to ANSI Color
}
\keyword{internal}
