% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shape.R
\name{merge_shapes}
\alias{merge_shapes}
\title{Merge shapes

Add shape to a mesh}
\usage{
merge_shapes(scene)
}
\arguments{
\item{scene}{}
}
\value{
Merged scene
}
\description{
Merge shapes

Add shape to a mesh
}
\keyword{internal}
