% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shape.R
\name{preprocess_scene}
\alias{preprocess_scene}
\title{Preprocess Scene}
\usage{
preprocess_scene(scene_list)
}
\arguments{
\item{scene_list}{The scene list.}
}
\value{
Scene list converted to proper format.
}
\description{
Preprocess the scene
}
\keyword{internal}
