% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_mesh.R
\name{rasterize_mesh}
\alias{rasterize_mesh}
\title{Rasterize an OBJ file}
\usage{
rasterize_mesh(
  mesh,
  filename = NA,
  width = 400,
  height = 400,
  line_info = NULL,
  alpha_line = 0.5,
  parallel = TRUE,
  fov = 20,
  lookfrom = c(0, 0, 10),
  lookat = NULL,
  camera_up = c(0, 1, 0),
  scale_obj = 1,
  light_info = directional_light(),
  color = "red",
  type = "diffuse",
  background = "white",
  tangent_space_normals = TRUE,
  ambient_color = c(0, 0, 0),
  exponent = 32,
  specular_intensity = 0.6,
  emission_intensity = 1,
  override_exponent = FALSE,
  diffuse_intensity = 1,
  shadow_map = FALSE,
  shadow_map_bias = 0.005,
  shadow_map_intensity = 0.5,
  shadow_map_dims = NULL,
  ssao = FALSE,
  ssao_intensity = 10,
  ssao_radius = 0.1,
  tonemap = "none",
  debug = "none",
  near_plane = 0.1,
  far_plane = 100,
  culling = "back",
  shader = "default",
  double_sided = FALSE,
  block_size = 4,
  shape = NULL,
  line_offset = -1e-05,
  ortho_dims = c(1, 1)
)
}
\arguments{
\item{shape}{Default `NULL`. The shape to render in the OBJ mesh.}

\item{obj_model}{A two-dimensional matrix, where each entry in the matrix is the elevation at that point. All points are assumed to be evenly spaced.}

\item{...}{Other parameters to pass to `rasterize_mesh()`.}
}
\value{
Rasterized image.
}
\description{
Rasterize an OBJ file
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
