% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_obj.R
\name{rasterize_obj}
\alias{rasterize_obj}
\title{Rasterize an OBJ file}
\usage{
rasterize_obj(
  obj_model,
  filename = NA,
  width = 400,
  height = 400,
  fov = 20,
  lookfrom = c(0, 0, 10),
  lookat = c(0, 0, 0),
  type = "diffuse",
  color = "darkred",
  texture_location = NA,
  normal_texture_location = NA,
  specular_texture_location = NA,
  light_direction = c(1, 1, 1),
  ambient_color = c(0, 0, 0),
  exponent = 32,
  specular_intensity = 0.6,
  diffuse_intensity = 1,
  tangent_space_normals = FALSE,
  shadow_map = FALSE
)
}
\arguments{
\item{obj_model}{A two-dimensional matrix, where each entry in the matrix is the elevation at that point. All points are assumed to be evenly spaced.}

\item{width}{Default `400`. Width of the rendered image.}

\item{height}{Default `400`. Width of the rendered image.}

\item{fov}{Default `20`. Width of the rendered image.}

\item{lookfrom}{Default `c(0,0,10)`. Camera location.}

\item{lookat}{Default `c(0,0,0)`. Camera focal position.}
}
\value{
Rasterized image.
}
\description{
Rasterize an OBJ file
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
