% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_obj.R
\name{rasterize_obj}
\alias{rasterize_obj}
\title{Rasterize an OBJ file}
\usage{
rasterize_obj(
  obj_model,
  filename = NA,
  width = 400,
  height = 400,
  fov = 20,
  lookfrom = c(0, 0, 10),
  lookat = NULL,
  camera_up = c(0, 1, 0),
  scale_obj = 1,
  point_light_info = NULL,
  type = "diffuse",
  color = "darkred",
  background = "white",
  texture_location = NA,
  normal_texture_location = NA,
  specular_texture_location = NA,
  emissive_texture_location = NA,
  parallel = TRUE,
  light_direction = c(1, 1, 1),
  light_intensity = 1,
  ambient_color = c(0, 0, 0),
  exponent = 32,
  specular_intensity = 0.6,
  emission_intensity = 1,
  override_exponent = FALSE,
  diffuse_intensity = 1,
  tangent_space_normals = TRUE,
  shadow_map = FALSE,
  calc_ambient = FALSE,
  ambient_intensity = 10,
  ambient_radius = 0.1,
  tonemap = "none",
  debug = "none",
  shadow_map_bias = -0.001,
  shadow_map_intensity = 0.5,
  near_plane = 0.1,
  far_plane = 100,
  culling = "back",
  shader = "default",
  double_sided = FALSE,
  block_size = 4,
  shape = NULL,
  shadow_map_dims = NULL
)
}
\arguments{
\item{obj_model}{A two-dimensional matrix, where each entry in the matrix is the elevation at that point. All points are assumed to be evenly spaced.}

\item{width}{Default `400`. Width of the rendered image.}

\item{height}{Default `400`. Width of the rendered image.}

\item{fov}{Default `20`. Width of the rendered image.}

\item{lookfrom}{Default `c(0,0,10)`. Camera location.}

\item{lookat}{Default `c(0,0,0)`. Camera focal position.}
}
\value{
Rasterized image.
}
\description{
Rasterize an OBJ file
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
