% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_obj.R
\name{rasterize_obj}
\alias{rasterize_obj}
\title{Rasterize an OBJ file

If the OBJ file has an MTL file, it must be located in the same directory as the OBJ file itself.
Additionally, the textures referenced in the MTL file must be specified as a 
relative path to the OBJ file's location.}
\usage{
rasterize_obj(
  obj_model,
  filename = NA,
  width = 400,
  height = 400,
  line_info = NULL,
  alpha_line = 0.5,
  parallel = TRUE,
  fov = 20,
  lookfrom = c(0, 0, 10),
  lookat = NULL,
  camera_up = c(0, 1, 0),
  scale_obj = 1,
  point_light_info = NULL,
  type = "diffuse",
  color = "darkred",
  background = "white",
  light_direction = c(1, 1, 1),
  light_intensity = 1,
  ambient_color = c(0, 0, 0),
  exponent = 32,
  specular_intensity = 0.6,
  emission_intensity = 1,
  override_exponent = FALSE,
  diffuse_intensity = 1,
  tangent_space_normals = TRUE,
  shadow_map = FALSE,
  shadow_map_bias = 0.005,
  shadow_map_intensity = 0.5,
  shadow_map_dims = NULL,
  ssao = FALSE,
  ssao_intensity = 10,
  ssao_radius = 0.1,
  tonemap = "none",
  debug = "none",
  near_plane = 0.1,
  far_plane = 100,
  culling = "back",
  shader = "default",
  double_sided = FALSE,
  block_size = 4,
  shape = NULL,
  line_offset = -1e-05
)
}
\arguments{
\item{obj_model}{Filename of the `obj` file.}

\item{filename}{Default `NULL`. Filename to save the image. If `NULL`, the image will be plotted.}

\item{width}{Default `400`. Width of the rendered image.}

\item{height}{Default `400`. Width of the rendered image.}

\item{line_info}{Default `NULL`. Matrix of line segments to add to the scene. Number of rows must be a multiple of 2.}

\item{parallel}{Default `TRUE`. Whether to use parallel processing.}

\item{fov}{Default `20`. Width of the rendered image.}

\item{lookfrom}{Default `c(0,0,10)`. Camera location.}

\item{lookat}{Default `NULL`. Camera focal position, defaults to the center of the model.}

\item{camera_up}{Default `c(0,1,0)`. Camera up vector.}

\item{scale_obj}{Default `1`. Value to scale size of model.}

\item{point_light_info}{Default `NULL`. A data.frame of point light information created using the 
`point_light()` and `add_light()` functions.}

\item{type}{Default `diffuse`. Shader type. Other options: `vertex` (Gouraud shading), `phong`, and `color` (no lighting).}

\item{color}{Default `darkred`. Color of model if no material file present (or for faces using the default material).}

\item{background}{Default `white`. Background color.}

\item{light_direction}{Default `c(1,1,1)`. Vector specifying the light direction for the primary directional light.}

\item{light_intensity}{Default `1.0`. Light intensity.}

\item{ambient_color}{Default `c(0,0,0)`. Ambient color of model if no material file present (or for faces using the default material).}

\item{exponent}{Default `32`.}

\item{specular_intensity}{Default `0.6`.}

\item{emission_intensity}{Default `1`.}

\item{override_exponent}{Default `FALSE`.}

\item{diffuse_intensity}{Default `1`.}

\item{tangent_space_normals}{Default `TRUE`.}

\item{shadow_map}{Default `FALSE`.}

\item{shadow_map_bias}{Default `0.005`.}

\item{shadow_map_intensity}{Default `0.5`.}

\item{shadow_map_dims}{Default `NULL`.}

\item{ssao}{Default `FALSE`. Whether to add screen-space ambient occlusion (SSAO) to the render.}

\item{ssao_intensity}{Default `10`. Intensity of the shadow map.}

\item{ssao_radius}{Default `0.1`. Radius to use when calculating the SSAO term.}

\item{tonemap}{Default `"none"`.}

\item{debug}{Default `"none"`.}

\item{near_plane}{Default `0.1`.}

\item{far_plane}{Default `100`.}

\item{culling}{Default `"back"`.}

\item{shader}{Default `"default"`.}

\item{block_size}{Default `4`.}

\item{shape}{Default `NULL`}
}
\value{
Rasterized image.
}
\description{
Rasterize an OBJ file

If the OBJ file has an MTL file, it must be located in the same directory as the OBJ file itself.
Additionally, the textures referenced in the MTL file must be specified as a 
relative path to the OBJ file's location.
}
\examples{
#Here we produce a ambient occlusion map of the `montereybay` elevation map.
}
