% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shape.R
\name{remove_duplicate_materials}
\alias{remove_duplicate_materials}
\title{Remove Duplicates}
\usage{
remove_duplicate_materials(scene)
}
\arguments{
\item{scene}{The scene}
}
\value{
Scene with shape added.
}
\description{
Remove Duplicates
}
\keyword{internal}
