% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shape.R
\name{remove_duplicate_materials}
\alias{remove_duplicate_materials}
\title{Add Shape

Add shape to the scene}
\usage{
remove_duplicate_materials(scene)
}
\arguments{
\item{scene}{The scene to add the shape.}
}
\value{
Scene with shape added.
}
\description{
Add Shape

Add shape to the scene
}
\keyword{#Generate}
\keyword{box}
\keyword{cornell}
\keyword{in}
\keyword{internal}
\keyword{several}
\keyword{spheres}
\keyword{the}
