% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{sphere_mesh}
\alias{sphere_mesh}
\title{Sphere 3D Model}
\usage{
sphere_mesh(
  position = c(0, 0, 0),
  radius = 1,
  scale = c(1, 1, 1),
  angle = c(0, 0, 0),
  pivot_point = c(0, 0, 0),
  order_rotation = c(1, 2, 3),
  low_poly = FALSE
)
}
\value{
File location of the R.obj file (saved with a .txt extension)
}
\description{
Sphere 3D Model
}
