% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_yz.R
\name{swap_yz}
\alias{swap_yz}
\title{Swap Y/Z Axis}
\usage{
swap_yz(mesh)
}
\arguments{
\item{mesh}{The mesh to swap orientations.}
}
\value{
Mesh with Y and Z axis exchanged
}
\description{
Swap Y/Z Axis
}
\examples{
# Flip a mesh 
sphere_mesh(position=c(-1,0,0)) |> 
  add_shape(swap_yz(sphere_mesh(position=c(0,1,0)))) |> 
  rasterize_scene()
}
